requirements.txt
geopandas=0.14.2
numpy=1.26.3
networkx=3.2.1
shapely=2.0.2
pytest=7.4.4
pre-commit
scipy=1.12.0
omegaconf=2.2.3
hydra-core=1.3.2
seaborn
momepy
igraph==0.10.2
statsmodels
community